+++ tests/TestCase/TestSuite/MockTraitTest.php
@@ -19,6 +19,8 @@ use App\Controller\PagesController;
 use App\Model\Table\PostsTable;
 use App\Model\Validation\PostValidator;
 use App\View\Cell\MyExampleCell;
+use Cake\Controller\Component;
+use Cake\Controller\Controller;
 use Cake\View\Helper;
 use MeTools\TestSuite\TestCase;
 use PHPUnit\Framework\AssertionFailedError;
@@ -45,37 +47,41 @@ class MockTraitTest extends TestCase
             \stdClass::class => 'Unable to get the alias for the `stdClass` class',
             'No\Existing\Class' => 'Class `No\Existing\Class` does not exist',
         ] as $className => $expectedMessage) {
-            $this->assertException(function () use ($className) {
-                /** @phpstan-ignore-next-line */
-                $this->getAlias($className);
-            }, AssertionFailedError::class, $expectedMessage);
+            /** @phpstan-ignore-next-line */
+            $this->assertException(fn() => $this->getAlias($className), AssertionFailedError::class, $expectedMessage);
         }
     }
 
     /**
      * Tests for `getMockForComponent()` method
      * @test
+     * @uses \MeTools\TestSuite\MockTrait::getMockForComponent()
      */
     public function testGetMockForComponent(): void
     {
-        $this->assertIsMock($this->getMockForComponent('Cake\Controller\Component\FlashComponent'));
+        $Component = $this->getMockForComponent('Cake\Controller\Component\FlashComponent');
+        $this->assertIsMock($Component);
+        $this->assertInstanceOf(Component::class, $Component);
     }
 
     /**
      * Tests for `getMockForComponent()` method
      * @test
+     * @uses \MeTools\TestSuite\MockTrait::getMockForController()
      */
     public function testGetMockForController(): void
     {
-        /** @var \App\Controller\PagesController $Mock **/
-        $Mock = $this->getMockForController('App\Controller\PagesController', null);
-        $this->assertIsMock($Mock);
-        $this->assertEquals('Pages', $Mock->getName());
+        /** @var \App\Controller\PagesController $Controller **/
+        $Controller = $this->getMockForController('App\Controller\PagesController', null);
+        $this->assertIsMock($Controller);
+        $this->assertInstanceOf(Controller::class, $Controller);
+        $this->assertEquals('Pages', $Controller->getName());
 
-        /** @var \App\Controller\PagesController $Mock **/
-        $Mock = $this->getMockForController('App\Controller\PagesController', null, 'MyController');
-        $this->assertIsMock($Mock);
-        $this->assertEquals('MyController', $Mock->getName());
+        /** @var \App\Controller\PagesController $Controller **/
+        $Controller = $this->getMockForController('App\Controller\PagesController', null, 'MyController');
+        $this->assertIsMock($Controller);
+        $this->assertInstanceOf(Controller::class, $Controller);
+        $this->assertEquals('MyController', $Controller->getName());
 
         //With a no existing class
         $this->expectAssertionFailed('Class `App\Controller\NoExistingController` does not exist');